The type of the 1st parameter of the Converter::money() method has been changed from int to float.

+++ Internal/package-helpers/Converter.php
@@ -206,13 +206,13 @@ class Converter
     /**
      * Convert Money
      * 
-     * @param int    $money = 0
+     * @param float  $money = 0
      * @param string $type  = NULL
      * @param bool   $float = true 
      * 
      * @return string
      */
-    public static function money(Int $money = 0, String $type = NULL, Bool $float = true) : String
+    public static function money(Float $money = 0, String $type = NULL, Bool $float = true) : String
     {
         $moneyFormat = '';
         $money       = round($money, 2);
